+++ frontend/src/components/common/BucketCard/BucketCard.interface.ts
@@ -1,4 +1,4 @@
-import { LadderBandBucket } from '@app/types/common';
+import { LadderBandBucket } from '@app/types/library';
 
 export interface BucketCardProps {
   bucket: LadderBandBucket;
