+++ frontend/src/components/modules/AdvancementLevel/AdvancemetLevel.interface.ts
@@ -1,4 +1,4 @@
-import { AdvancementLevel } from '@app/types/common';
+import { AdvancementLevel } from '@app/types/library';
 
 export interface AdvancementLevelProps {
   data: AdvancementLevel;
