+++ frontend/src/components/modules/BucketDetails/BucketDetails.interface.ts
@@ -1,4 +1,4 @@
-import { Bucket } from '@app/types/common';
+import { Bucket } from '@app/types/library';
 
 export interface BucketDetailsProps {
   data: Bucket;
