+++ frontend/src/components/modules/LadderDetails/LadderDetails.interface.ts
@@ -1,4 +1,4 @@
-import { LadderBand } from '@app/types/common';
+import { LadderBand } from '@app/types/library';
 
 export interface LadderDetailsProps {
   ladder: LadderBand;
